pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pytest_reserial"
authors = [{name = "Alexander Bessman", email = "alexander.bessman@gmail.com"}]
dynamic = ["version", "description"]
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
dependencies = [
    "pytest",
    "pyserial",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
]
keywords = [
    "serial",
    "testing",
    "logging",
]

[project.urls]
Home = "https://github.com/bessman/pytest-reserial"

[project.entry-points."pytest11"]
reserial = "pytest_reserial.reserial"